+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(OpenSketch)
-
-test_check("OpenSketch")